Services module refactored
